+++ ros_ign_bridge/src/convert/frame_id.cpp
@@ -1,0 +1,56 @@
+// Copyright 2021 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "convert/frame_id.hpp"
+
+namespace ros_ign_bridge
+{
+// This can be used to replace `::` with `/` to make frame_id compatible with TF
+std::string replace_delimiter(const std::string &input,
+                              const std::string &old_delim,
+                              const std::string new_delim)
+{
+  std::string output;
+  output.reserve(input.size());
+
+  std::size_t last_pos = 0;
+
+  while (last_pos < input.size())
+  {
+    std::size_t pos = input.find(old_delim, last_pos);
+    output += input.substr(last_pos, pos - last_pos);
+    if (pos != std::string::npos)
+    {
+      output += new_delim;
+      pos += old_delim.size();
+    }
+
+    last_pos = pos;
+  }
+
+  return output;
+}
+
+// Frame id from ROS to ign is not supported right now
+// std::string frame_id_ros_to_ign(const std::string &frame_id)
+// {
+//   return replace_delimiter(frame_id, "/", "::");
+// }
+
+std::string frame_id_ign_to_ros(const std::string &frame_id)
+{
+  return replace_delimiter(frame_id, "::", "/");
+}
+
+}  // namespace ros_ign_bridge
