+++ ros_ign_bridge/src/convert/sensor_msgs.cpp
@@ -1,4 +1,4 @@
-// Copyright 2018 Open Source Robotics Foundation, Inc.
+// Copyright 2021 Open Source Robotics Foundation, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,604 +17,10 @@
 #include <ros/console.h>
 
 #include "ros_ign_bridge/convert.hpp"
+#include "convert/frame_id.hpp"
 
 namespace ros_ign_bridge
 {
-
-// This can be used to replace `::` with `/` to make frame_id compatible with TF
-std::string replace_delimiter(const std::string &input,
-                              const std::string &old_delim,
-                              const std::string new_delim)
-{
-  std::string output;
-  output.reserve(input.size());
-
-  std::size_t last_pos = 0;
-
-  while (last_pos < input.size())
-  {
-    std::size_t pos = input.find(old_delim, last_pos);
-    output += input.substr(last_pos, pos - last_pos);
-    if (pos != std::string::npos)
-    {
-      output += new_delim;
-      pos += old_delim.size();
-    }
-
-    last_pos = pos;
-  }
-
-  return output;
-}
-
-// Frame id from ROS to ign is not supported right now
-// std::string frame_id_ros_to_ign(const std::string &frame_id)
-// {
-//   return replace_delimiter(frame_id, "/", "::");
-// }
-
-std::string frame_id_ign_to_ros(const std::string &frame_id)
-{
-  return replace_delimiter(frame_id, "::", "/");
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Bool & ros_msg,
-  ignition::msgs::Boolean & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Boolean & ign_msg,
-  std_msgs::Bool & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::ColorRGBA & ros_msg,
-  ignition::msgs::Color & ign_msg)
-{
-  ign_msg.set_r(ros_msg.r);
-  ign_msg.set_g(ros_msg.g);
-  ign_msg.set_b(ros_msg.b);
-  ign_msg.set_a(ros_msg.a);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Color & ign_msg,
-  std_msgs::ColorRGBA & ros_msg)
-{
-  ros_msg.r = ign_msg.r();
-  ros_msg.g = ign_msg.g();
-  ros_msg.b = ign_msg.b();
-  ros_msg.a = ign_msg.a();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Empty &,
-  ignition::msgs::Empty &)
-{
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Empty &,
-  std_msgs::Empty &)
-{
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Int32 & ros_msg,
-  ignition::msgs::Int32 & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Int32 & ign_msg,
-  std_msgs::Int32 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Float32 & ros_msg,
-  ignition::msgs::Float & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Float & ign_msg,
-  std_msgs::Float32 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Float64 & ros_msg,
-  ignition::msgs::Double & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Double & ign_msg,
-  std_msgs::Float64 & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::Header & ros_msg,
-  ignition::msgs::Header & ign_msg)
-{
-  ign_msg.mutable_stamp()->set_sec(ros_msg.stamp.sec);
-  ign_msg.mutable_stamp()->set_nsec(ros_msg.stamp.nsec);
-  auto newPair = ign_msg.add_data();
-  newPair->set_key("seq");
-  newPair->add_value(std::to_string(ros_msg.seq));
-  newPair = ign_msg.add_data();
-  newPair->set_key("frame_id");
-  newPair->add_value(ros_msg.frame_id);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Header & ign_msg,
-  std_msgs::Header & ros_msg)
-{
-  ros_msg.stamp = ros::Time(ign_msg.stamp().sec(), ign_msg.stamp().nsec());
-  for (auto i = 0; i < ign_msg.data_size(); ++i)
-  {
-    auto aPair = ign_msg.data(i);
-    if (aPair.key() == "seq" && aPair.value_size() > 0)
-    {
-      std::string value = aPair.value(0);
-      try
-      {
-        unsigned long ul = std::stoul(value, nullptr);
-        ros_msg.seq = ul;
-      }
-      catch (std::exception & e)
-      {
-        ROS_ERROR_STREAM("Exception converting [" << value << "] to an "
-                  << "unsigned int" << std::endl);
-      }
-    }
-    else if (aPair.key() == "frame_id" && aPair.value_size() > 0)
-    {
-      ros_msg.frame_id = frame_id_ign_to_ros(aPair.value(0));
-    }
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const std_msgs::String & ros_msg,
-  ignition::msgs::StringMsg & ign_msg)
-{
-  ign_msg.set_data(ros_msg.data);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::StringMsg & ign_msg,
-  std_msgs::String & ros_msg)
-{
-  ros_msg.data = ign_msg.data();
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Clock & ign_msg,
-  rosgraph_msgs::Clock & ros_msg)
-{
-  ros_msg.clock = ros::Time(ign_msg.sim().sec(), ign_msg.sim().nsec());
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const rosgraph_msgs::Clock & ros_msg,
-  ignition::msgs::Clock & ign_msg)
-{
-  ign_msg.mutable_sim()->set_sec(ros_msg.clock.sec);
-  ign_msg.mutable_sim()->set_nsec(ros_msg.clock.nsec);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Quaternion & ros_msg,
-  ignition::msgs::Quaternion & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-  ign_msg.set_w(ros_msg.w);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Quaternion & ign_msg,
-  geometry_msgs::Quaternion & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-  ros_msg.w = ign_msg.w();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Vector3 & ros_msg,
-  ignition::msgs::Vector3d & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Vector3d & ign_msg,
-  geometry_msgs::Vector3 & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Point & ros_msg,
-  ignition::msgs::Vector3d & ign_msg)
-{
-  ign_msg.set_x(ros_msg.x);
-  ign_msg.set_y(ros_msg.y);
-  ign_msg.set_z(ros_msg.z);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Vector3d & ign_msg,
-  geometry_msgs::Point & ros_msg)
-{
-  ros_msg.x = ign_msg.x();
-  ros_msg.y = ign_msg.y();
-  ros_msg.z = ign_msg.z();
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Pose & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.position, *ign_msg.mutable_position());
-  convert_ros_to_ign(ros_msg.orientation, *ign_msg.mutable_orientation());
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::Pose & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.position(), ros_msg.position);
-  convert_ign_to_ros(ign_msg.orientation(), ros_msg.orientation);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::PoseArray & ros_msg,
-  ignition::msgs::Pose_V & ign_msg)
-{
-  ign_msg.clear_pose();
-  for (auto const &t : ros_msg.poses)
-  {
-    auto p = ign_msg.add_pose();
-    convert_ros_to_ign(t, *p);
-  }
-
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose_V & ign_msg,
-  geometry_msgs::PoseArray & ros_msg)
-{
-  ros_msg.poses.clear();
-  for (auto const &p : ign_msg.pose())
-  {
-    geometry_msgs::Pose pose;
-    convert_ign_to_ros(p, pose);
-    ros_msg.poses.push_back(pose);
-  }
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::PoseStamped & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.pose, ign_msg);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::PoseStamped & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg, ros_msg.pose);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Transform & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.translation , *ign_msg.mutable_position());
-  convert_ros_to_ign(ros_msg.rotation, *ign_msg.mutable_orientation());
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::Transform & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.position(), ros_msg.translation);
-  convert_ign_to_ros(ign_msg.orientation(), ros_msg.rotation);
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::TransformStamped & ros_msg,
-  ignition::msgs::Pose & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.transform, ign_msg);
-
-  auto newPair = ign_msg.mutable_header()->add_data();
-  newPair->set_key("child_frame_id");
-  newPair->add_value(ros_msg.child_frame_id);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose & ign_msg,
-  geometry_msgs::TransformStamped & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg, ros_msg.transform);
-  for (auto i = 0; i < ign_msg.header().data_size(); ++i)
-  {
-    auto aPair = ign_msg.header().data(i);
-    if (aPair.key() == "child_frame_id" && aPair.value_size() > 0)
-    {
-      ros_msg.child_frame_id = frame_id_ign_to_ros(aPair.value(0));
-      break;
-    }
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const tf2_msgs::TFMessage & ros_msg,
-  ignition::msgs::Pose_V & ign_msg)
-{
-  ign_msg.clear_pose();
-  for (auto const &t : ros_msg.transforms)
-  {
-    auto p = ign_msg.add_pose();
-    convert_ros_to_ign(t, *p);
-  }
-
-  if (!ros_msg.transforms.empty())
-  {
-    convert_ros_to_ign(ros_msg.transforms[0].header,
-        (*ign_msg.mutable_header()));
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Pose_V & ign_msg,
-  tf2_msgs::TFMessage & ros_msg)
-{
-  ros_msg.transforms.clear();
-  for (auto const &p : ign_msg.pose())
-  {
-    geometry_msgs::TransformStamped tf;
-    convert_ign_to_ros(p, tf);
-    ros_msg.transforms.push_back(tf);
-  }
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const geometry_msgs::Twist & ros_msg,
-  ignition::msgs::Twist & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.linear,  (*ign_msg.mutable_linear()));
-  convert_ros_to_ign(ros_msg.angular, (*ign_msg.mutable_angular()));
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Twist & ign_msg,
-  geometry_msgs::Twist & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.linear(), ros_msg.linear);
-  convert_ign_to_ros(ign_msg.angular(), ros_msg.angular);
-}
-
-// template<>
-// void
-// convert_ros_to_ign(
-//   const mav_msgs::Actuators & ros_msg,
-//   ignition::msgs::Actuators & ign_msg)
-// {
-//   convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-//
-//   for (auto i = 0u; i < ros_msg.angles.size(); ++i)
-//     ign_msg.add_position(ros_msg.angles[i]);
-//   for (auto i = 0u; i < ros_msg.angular_velocities.size(); ++i)
-//     ign_msg.add_velocity(ros_msg.angular_velocities[i]);
-//   for (auto i = 0u; i < ros_msg.normalized.size(); ++i)
-//     ign_msg.add_normalized(ros_msg.normalized[i]);
-// }
-//
-// template<>
-// void
-// convert_ign_to_ros(
-//   const ignition::msgs::Actuators & ign_msg,
-//   mav_msgs::Actuators & ros_msg)
-// {
-//   convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-//
-//   for (auto i = 0; i < ign_msg.position_size(); ++i)
-//     ros_msg.angles.push_back(ign_msg.position(i));
-//   for (auto i = 0; i < ign_msg.velocity_size(); ++i)
-//     ros_msg.angular_velocities.push_back(ign_msg.velocity(i));
-//   for (auto i = 0; i < ign_msg.normalized_size(); ++i)
-//     ros_msg.normalized.push_back(ign_msg.normalized(i));
-// }
-
-template<>
-void
-convert_ros_to_ign(
-  const nav_msgs::OccupancyGrid & ros_msg,
-  ignition::msgs::OccupancyGrid & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-
-  ign_msg.mutable_info()->mutable_map_load_time()->set_sec(
-      ros_msg.info.map_load_time.sec);
-  ign_msg.mutable_info()->mutable_map_load_time()->set_nsec(
-      ros_msg.info.map_load_time.nsec);
-
-  ign_msg.mutable_info()->set_resolution(
-      ros_msg.info.resolution);
-  ign_msg.mutable_info()->set_width(
-      ros_msg.info.width);
-  ign_msg.mutable_info()->set_height(
-      ros_msg.info.height);
-
-  convert_ros_to_ign(ros_msg.info.origin, 
-      (*ign_msg.mutable_info()->mutable_origin()));
-
-  ign_msg.set_data(&ros_msg.data[0], ros_msg.data.size());
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::OccupancyGrid & ign_msg,
-  nav_msgs::OccupancyGrid & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-
-  ros_msg.info.map_load_time.sec = ign_msg.info().map_load_time().sec();
-  ros_msg.info.map_load_time.nsec = ign_msg.info().map_load_time().nsec();
-  ros_msg.info.resolution = ign_msg.info().resolution();
-  ros_msg.info.width = ign_msg.info().width();
-  ros_msg.info.height = ign_msg.info().height();
-
-  convert_ign_to_ros(ign_msg.info().origin(), ros_msg.info.origin);
-
-  ros_msg.data.resize(ign_msg.data().size());
-  memcpy(&ros_msg.data[0], ign_msg.data().c_str(), ign_msg.data().size());
-}
-
-template<>
-void
-convert_ros_to_ign(
-  const nav_msgs::Odometry & ros_msg,
-  ignition::msgs::Odometry & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-  convert_ros_to_ign(ros_msg.pose.pose, (*ign_msg.mutable_pose()));
-  convert_ros_to_ign(ros_msg.twist.twist, (*ign_msg.mutable_twist()));
-
-  auto childFrame = ign_msg.mutable_header()->add_data();
-  childFrame->set_key("child_frame_id");
-  childFrame->add_value(ros_msg.child_frame_id);
-}
-
-template<>
-void
-convert_ign_to_ros(
-  const ignition::msgs::Odometry & ign_msg,
-  nav_msgs::Odometry & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-  convert_ign_to_ros(ign_msg.pose(), ros_msg.pose.pose);
-  convert_ign_to_ros(ign_msg.twist(), ros_msg.twist.twist);
-
-  for (auto i = 0; i < ign_msg.header().data_size(); ++i)
-  {
-    auto aPair = ign_msg.header().data(i);
-    if (aPair.key() == "child_frame_id" && aPair.value_size() > 0)
-    {
-      ros_msg.child_frame_id = frame_id_ign_to_ros(aPair.value(0));
-      break;
-    }
-  }
-}
 
 template<>
 void
@@ -1345,258 +751,5 @@ convert_ign_to_ros(
   ros_msg.present = true;
 }
 
-template<>
-void
-convert_ros_to_ign(
-    const visualization_msgs::Marker & ros_msg,
-    ignition::msgs::Marker & ign_msg)
-{
-  convert_ros_to_ign(ros_msg.header, (*ign_msg.mutable_header()));
-
-  // Note, in ROS's Marker message ADD and MODIFY both map to a value of "0", 
-  // so that case is not needed here.
-  switch(ros_msg.action)
-  {
-    case visualization_msgs::Marker::ADD:
-      ign_msg.set_action(ignition::msgs::Marker::ADD_MODIFY);
-      break;
-    case visualization_msgs::Marker::DELETE:
-      ign_msg.set_action(ignition::msgs::Marker::DELETE_MARKER);
-      break;
-    case visualization_msgs::Marker::DELETEALL:
-      ign_msg.set_action(ignition::msgs::Marker::DELETE_ALL);
-      break;
-    default:
-      ROS_ERROR_STREAM("Unknown visualization_msgs::Marker action [" <<
-          ros_msg.action << "]\n");
-      break;
-  }
-
-  ign_msg.set_ns(ros_msg.ns);
-  ign_msg.set_id(ros_msg.id);
-  // ign_msg.set_layer();  // No "layer" concept in ROS
-
-  // Type
-  switch(ros_msg.type)
-  {
-#ifdef IGNITION_DOME
-    case visualization_msgs::Marker::ARROW:
-      ign_msg.set_type(ignition::msgs::Marker::ARROW);
-      break;
-#endif
-    case visualization_msgs::Marker::CUBE:
-      ign_msg.set_type(ignition::msgs::Marker::BOX);
-      break;
-    case visualization_msgs::Marker::SPHERE:
-      ign_msg.set_type(ignition::msgs::Marker::SPHERE);
-      break;
-    case visualization_msgs::Marker::CYLINDER:
-      ign_msg.set_type(ignition::msgs::Marker::CYLINDER);
-      break;
-    case visualization_msgs::Marker::LINE_STRIP:
-      ign_msg.set_type(ignition::msgs::Marker::LINE_STRIP);
-      break;
-    case visualization_msgs::Marker::LINE_LIST:
-      ign_msg.set_type(ignition::msgs::Marker::LINE_LIST);
-      break;
-    case visualization_msgs::Marker::CUBE_LIST:
-      ROS_ERROR_STREAM("Unsupported visualization_msgs::Marker type" <<
-          "[CUBE_LIST]\n");
-      break;
-    case visualization_msgs::Marker::SPHERE_LIST:
-      ROS_ERROR_STREAM("Unsupported visualization_msgs::Marker type" <<
-          "[SPHERE_LIST]\n");
-      break;
-    case visualization_msgs::Marker::POINTS:
-      ign_msg.set_type(ignition::msgs::Marker::POINTS);
-      break;
-    case visualization_msgs::Marker::TEXT_VIEW_FACING:
-      ign_msg.set_type(ignition::msgs::Marker::TEXT);
-      break;
-    case visualization_msgs::Marker::MESH_RESOURCE:
-      ROS_ERROR_STREAM("Unsupported visualization_msgs::Marker type" <<
-          "[MESH_RESOURCE]\n");
-      break;
-    case visualization_msgs::Marker::TRIANGLE_LIST:
-      ign_msg.set_type(ignition::msgs::Marker::TRIANGLE_LIST);
-      break;
-    default:
-      ROS_ERROR_STREAM("Unknown visualization_msgs::Marker type [" <<
-          ros_msg.type << "]\n");
-      break;
-  }
-
-  // Lifetime
-  ign_msg.mutable_lifetime()->set_sec(ros_msg.lifetime.sec);
-  ign_msg.mutable_lifetime()->set_nsec(ros_msg.lifetime.nsec);
-
-  // Pose
-  convert_ros_to_ign(ros_msg.pose, *ign_msg.mutable_pose());
-
-  // Scale
-  convert_ros_to_ign(ros_msg.scale, *ign_msg.mutable_scale());
-
-  // Material
-  convert_ros_to_ign(ros_msg.color, *ign_msg.mutable_material()->mutable_ambient());
-  convert_ros_to_ign(ros_msg.color, *ign_msg.mutable_material()->mutable_diffuse());
-  convert_ros_to_ign(ros_msg.color, *ign_msg.mutable_material()->mutable_specular());
-
-  // Point
-  ign_msg.clear_point();
-  for (auto const &pt : ros_msg.points)
-  {
-    auto p = ign_msg.add_point();
-    convert_ros_to_ign(pt, *p);
-  }
-
-  ign_msg.set_text(ros_msg.text);
-
-  // ign_msg.set_parent();  // No "parent" concept in ROS
-  // ign_msg.set_visibility();  // No "visibility" concept in ROS
-}
-
-template<>
-void
-convert_ign_to_ros(
-    const ignition::msgs::Marker & ign_msg,
-    visualization_msgs::Marker & ros_msg)
-{
-  convert_ign_to_ros(ign_msg.header(), ros_msg.header);
-
-  switch(ign_msg.action())
-  {
-    case ignition::msgs::Marker::ADD_MODIFY:
-      ros_msg.action = visualization_msgs::Marker::ADD;
-      break;
-    case ignition::msgs::Marker::DELETE_MARKER:
-      ros_msg.action = visualization_msgs::Marker::DELETE;
-      break;
-    case ignition::msgs::Marker::DELETE_ALL:
-      ros_msg.action = visualization_msgs::Marker::DELETEALL;
-      break;
-    default:
-      ROS_ERROR_STREAM("Unknown ignition.msgs.marker action [" <<
-          ign_msg.action() << "]\n");
-      break;
-  }
-
-  ros_msg.ns = ign_msg.ns();
-  ros_msg.id = ign_msg.id();
-
-  switch(ign_msg.type())
-  {
-#ifdef IGNITION_DOME
-    case ignition::msgs::Marker::ARROW:
-      ros_msg.type = visualization_msgs::Marker::TRIANGLE_LIST;
-      break;
-    case ignition::msgs::Marker::AXIS:
-      ROS_ERROR_STREAM("Unsupported ignition.msgs.marker type " <<
-          "[AXIS]\n");
-      break;
-    case ignition::msgs::Marker::CONE:
-      ROS_ERROR_STREAM("Unsupported ignition.msgs.marker type " <<
-          "[CONE]\n");
-      break;
-#endif
-    case ignition::msgs::Marker::NONE:
-      ROS_ERROR_STREAM("Unsupported ignition.msgs.marker type " <<
-          "[NONE]\n");
-      break;
-    case ignition::msgs::Marker::BOX:
-      ros_msg.type = visualization_msgs::Marker::CUBE;
-      break;
-    case ignition::msgs::Marker::CYLINDER:
-      ros_msg.type = visualization_msgs::Marker::CYLINDER;
-      break;
-    case ignition::msgs::Marker::LINE_LIST:
-      ros_msg.type = visualization_msgs::Marker::LINE_LIST;
-      break;
-    case ignition::msgs::Marker::LINE_STRIP:
-      ros_msg.type = visualization_msgs::Marker::LINE_STRIP;
-      break;
-    case ignition::msgs::Marker::POINTS:
-      ros_msg.type = visualization_msgs::Marker::POINTS;
-      break;
-    case ignition::msgs::Marker::SPHERE:
-      ros_msg.type = visualization_msgs::Marker::SPHERE;
-      break;
-    case ignition::msgs::Marker::TEXT:
-      ros_msg.type = visualization_msgs::Marker::TEXT_VIEW_FACING;
-      break;
-    case ignition::msgs::Marker::TRIANGLE_FAN:
-      ROS_ERROR_STREAM("Unsupported ignition.msgs.marker type " <<
-          "[TRIANGLE_FAN]\n");
-      break;
-    case ignition::msgs::Marker::TRIANGLE_LIST:
-      ros_msg.type = visualization_msgs::Marker::TRIANGLE_LIST;
-      break;
-    case ignition::msgs::Marker::TRIANGLE_STRIP:
-      ROS_ERROR_STREAM("Unsupported ignition.msgs.marker type " <<
-          "[TRIANGLE_STRIP]\n");
-      break;
-    default:
-      ROS_ERROR_STREAM("Unknown ignition.msgs.marker type " <<
-          "[" << ign_msg.type() << "]\n");
-      break;
-  }
-
-  // Lifetime
-  ros_msg.lifetime.sec = ign_msg.lifetime().sec();
-  ros_msg.lifetime.nsec = ign_msg.lifetime().nsec();
-
-  // Pose
-  convert_ign_to_ros(ign_msg.pose(), ros_msg.pose);
-
-  // Scale
-  convert_ign_to_ros(ign_msg.scale(), ros_msg.scale);
-
-  // Material
-  convert_ign_to_ros(ign_msg.material().ambient(), ros_msg.color);
-  ros_msg.colors.clear();
-
-  // Points
-  ros_msg.points.clear();
-  ros_msg.points.reserve(ign_msg.point_size());
-  for (auto const & pt: ign_msg.point())
-  {
-    geometry_msgs::Point p;
-    convert_ign_to_ros(pt, p);
-    ros_msg.points.push_back(p);
-  }
-
-  ros_msg.text = ign_msg.text();
-}
-
-template<>
-void
-convert_ros_to_ign(
-    const visualization_msgs::MarkerArray & ros_msg,
-    ignition::msgs::Marker_V & ign_msg)
-{
-  ign_msg.clear_header();
-  ign_msg.clear_marker();
-  for (const auto &marker : ros_msg.markers)
-  {
-    auto m = ign_msg.add_marker();
-    convert_ros_to_ign(marker, *m);
-  }
-}
-
-template<>
-void
-convert_ign_to_ros(
-    const ignition::msgs::Marker_V & ign_msg,
-    visualization_msgs::MarkerArray & ros_msg)
-{
-  ros_msg.markers.clear();
-  ros_msg.markers.reserve(ign_msg.marker_size());
-
-  for (auto const &marker : ign_msg.marker())
-  {
-      visualization_msgs::Marker m;
-      convert_ign_to_ros(marker, m);
-      ros_msg.markers.push_back(m);
-  }
-}
-
 }  // namespace ros_ign_bridge
+
