+++ ros_ign_bridge/src/factories/geometry_msgs.cpp
@@ -1,0 +1,392 @@
+// Copyright 2018 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// geometry_msgs
+
+#include "ros_ign_bridge/convert.hpp"
+#include "factories/geometry_msgs.hpp"
+
+namespace ros_ign_bridge
+{
+
+std::shared_ptr<FactoryInterface>
+get_factory_geometry_msgs(
+  const std::string & ros_type_name,
+  const std::string & ign_type_name)
+{
+  if (
+    (ros_type_name == "geometry_msgs/Quaternion" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Quaternion")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Quaternion,
+        ignition::msgs::Quaternion
+      >
+    >("geometry_msgs/Quaternion", ign_type_name);
+  }
+  if (
+    (ros_type_name == "rosgraph_msgs/Clock" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Clock")
+  {
+    return std::make_shared<
+      Factory<
+        rosgraph_msgs::Clock,
+        ignition::msgs::Clock
+      >
+    >("rosgraph_msgs/Clock", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/Vector3" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Vector3d")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Vector3,
+        ignition::msgs::Vector3d
+      >
+    >("geometry_msgs/Vector3", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/Point" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Vector3d")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Point,
+        ignition::msgs::Vector3d
+      >
+    >("geometry_msgs/Point", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/Pose" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Pose,
+        ignition::msgs::Pose
+      >
+    >("geometry_msgs/Pose", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/PoseArray" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose_V")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::PoseArray,
+        ignition::msgs::Pose_V
+      >
+    >("geometry_msgs/PoseArray", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/PoseStamped" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::PoseStamped,
+        ignition::msgs::Pose
+      >
+    >("geometry_msgs/PoseStamped", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/Transform" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Transform,
+        ignition::msgs::Pose
+      >
+    >("geometry_msgs/Transform", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/TransformStamped" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::TransformStamped,
+        ignition::msgs::Pose
+      >
+    >("geometry_msgs/TransformStamped", ign_type_name);
+  }
+  if (
+    (ros_type_name == "tf2_msgs/TFMessage" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Pose_V")
+  {
+    return std::make_shared<
+      Factory<
+        tf2_msgs::TFMessage,
+        ignition::msgs::Pose_V
+      >
+    >("tf2_msgs/TFMessage", ign_type_name);
+  }
+  if (
+    (ros_type_name == "geometry_msgs/Twist" || ros_type_name == "") &&
+     ign_type_name == "ignition.msgs.Twist")
+  {
+    return std::make_shared<
+      Factory<
+        geometry_msgs::Twist,
+        ignition::msgs::Twist
+      >
+    >("geometry_msgs/Twist", ign_type_name);
+  }
+
+  return nullptr;
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Quaternion,
+  ignition::msgs::Quaternion
+>::convert_ros_to_ign(
+  const geometry_msgs::Quaternion & ros_msg,
+  ignition::msgs::Quaternion & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Quaternion,
+  ignition::msgs::Quaternion
+>::convert_ign_to_ros(
+  const ignition::msgs::Quaternion & ign_msg,
+  geometry_msgs::Quaternion & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Vector3,
+  ignition::msgs::Vector3d
+>::convert_ros_to_ign(
+  const geometry_msgs::Vector3 & ros_msg,
+  ignition::msgs::Vector3d & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Vector3,
+  ignition::msgs::Vector3d
+>::convert_ign_to_ros(
+  const ignition::msgs::Vector3d & ign_msg,
+  geometry_msgs::Vector3 & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Point,
+  ignition::msgs::Vector3d
+>::convert_ros_to_ign(
+  const geometry_msgs::Point & ros_msg,
+  ignition::msgs::Vector3d & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Point,
+  ignition::msgs::Vector3d
+>::convert_ign_to_ros(
+  const ignition::msgs::Vector3d & ign_msg,
+  geometry_msgs::Point & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Pose,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::Pose & ros_msg,
+  ignition::msgs::Pose & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Pose,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::Pose & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseArray,
+  ignition::msgs::Pose_V
+>::convert_ros_to_ign(
+  const geometry_msgs::PoseArray & ros_msg,
+  ignition::msgs::Pose_V & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseArray,
+  ignition::msgs::Pose_V
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose_V & ign_msg,
+  geometry_msgs::PoseArray & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseStamped,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::PoseStamped & ros_msg,
+  ignition::msgs::Pose & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseStamped,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::PoseStamped & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Transform,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::Transform & ros_msg,
+  ignition::msgs::Pose & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Transform,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::Transform & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::TransformStamped,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::TransformStamped & ros_msg,
+  ignition::msgs::Pose & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::TransformStamped,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::TransformStamped & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  tf2_msgs::TFMessage,
+  ignition::msgs::Pose_V
+>::convert_ros_to_ign(
+  const tf2_msgs::TFMessage & ros_msg,
+  ignition::msgs::Pose_V & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  tf2_msgs::TFMessage,
+  ignition::msgs::Pose_V
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose_V & ign_msg,
+  tf2_msgs::TFMessage & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Twist,
+  ignition::msgs::Twist
+>::convert_ros_to_ign(
+  const geometry_msgs::Twist & ros_msg,
+  ignition::msgs::Twist & ign_msg)
+{
+  ros_ign_bridge::convert_ros_to_ign(ros_msg, ign_msg);
+}
+
+template<>
+void
+Factory<
+  geometry_msgs::Twist,
+  ignition::msgs::Twist
+>::convert_ign_to_ros(
+  const ignition::msgs::Twist & ign_msg,
+  geometry_msgs::Twist & ros_msg)
+{
+  ros_ign_bridge::convert_ign_to_ros(ign_msg, ros_msg);
+}
+
+}  // namespace ros_ign_bridge
