+++ ros_ign_bridge/src/factories/geometry_msgs.hpp
@@ -1,0 +1,221 @@
+// Copyright 2021 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef ROS_IGN_BRIDGE__FACTORIES__GEOMETRY_MSGS_HPP_
+#define ROS_IGN_BRIDGE__FACTORIES__GEOMETRY_MSGS_HPP_
+
+// ROS messages
+#include <geometry_msgs/Quaternion.h>
+#include <geometry_msgs/Point.h>
+#include <geometry_msgs/Pose.h>
+#include <geometry_msgs/PoseArray.h>
+#include <geometry_msgs/PoseStamped.h>
+#include <geometry_msgs/Transform.h>
+#include <geometry_msgs/TransformStamped.h>
+#include <geometry_msgs/Twist.h>
+#include <geometry_msgs/Vector3.h>
+#include <tf2_msgs/TFMessage.h>
+
+#include "factory.hpp"
+
+namespace ros_ign_bridge 
+{
+
+std::shared_ptr<FactoryInterface>
+get_factory_geometry_msgs(
+            const std::string & ros_type_name,
+            const std::string & ign_type_name);
+
+template<>
+void
+Factory<
+  geometry_msgs::Quaternion,
+  ignition::msgs::Quaternion
+>::convert_ros_to_ign(
+  const geometry_msgs::Quaternion & ros_msg,
+  ignition::msgs::Quaternion & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Quaternion,
+  ignition::msgs::Quaternion
+>::convert_ign_to_ros(
+  const ignition::msgs::Quaternion & ign_msg,
+  geometry_msgs::Quaternion & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Vector3,
+  ignition::msgs::Vector3d
+>::convert_ros_to_ign(
+  const geometry_msgs::Vector3 & ros_msg,
+  ignition::msgs::Vector3d & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Vector3,
+  ignition::msgs::Vector3d
+>::convert_ign_to_ros(
+  const ignition::msgs::Vector3d & ign_msg,
+  geometry_msgs::Vector3 & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Point,
+  ignition::msgs::Vector3d
+>::convert_ros_to_ign(
+  const geometry_msgs::Point & ros_msg,
+  ignition::msgs::Vector3d & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Point,
+  ignition::msgs::Vector3d
+>::convert_ign_to_ros(
+  const ignition::msgs::Vector3d & ign_msg,
+  geometry_msgs::Point & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Pose,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::Pose & ros_msg,
+  ignition::msgs::Pose & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Pose,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::Pose & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseArray,
+  ignition::msgs::Pose_V
+>::convert_ros_to_ign(
+  const geometry_msgs::PoseArray & ros_msg,
+  ignition::msgs::Pose_V & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseArray,
+  ignition::msgs::Pose_V
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose_V & ign_msg,
+  geometry_msgs::PoseArray & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseStamped,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::PoseStamped & ros_msg,
+  ignition::msgs::Pose & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::PoseStamped,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::PoseStamped & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Transform,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::Transform & ros_msg,
+  ignition::msgs::Pose & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Transform,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::Transform & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::TransformStamped,
+  ignition::msgs::Pose
+>::convert_ros_to_ign(
+  const geometry_msgs::TransformStamped & ros_msg,
+  ignition::msgs::Pose & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::TransformStamped,
+  ignition::msgs::Pose
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose & ign_msg,
+  geometry_msgs::TransformStamped & ros_msg);
+
+template<>
+void
+Factory<
+  tf2_msgs::TFMessage,
+  ignition::msgs::Pose_V
+>::convert_ros_to_ign(
+  const tf2_msgs::TFMessage & ros_msg,
+  ignition::msgs::Pose_V & ign_msg);
+
+template<>
+void
+Factory<
+  tf2_msgs::TFMessage,
+  ignition::msgs::Pose_V
+>::convert_ign_to_ros(
+  const ignition::msgs::Pose_V & ign_msg,
+  tf2_msgs::TFMessage & ros_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Twist,
+  ignition::msgs::Twist
+>::convert_ros_to_ign(
+  const geometry_msgs::Twist & ros_msg,
+  ignition::msgs::Twist & ign_msg);
+
+template<>
+void
+Factory<
+  geometry_msgs::Twist,
+  ignition::msgs::Twist
+>::convert_ign_to_ros(
+  const ignition::msgs::Twist & ign_msg,
+  geometry_msgs::Twist & ros_msg);
+
+}  // namespace ros_ign_bridge
+#endif  // ROS_IGN_BRIDGE__FACTORIES__GEOMETRY_MSGS_HPP_
